Update: delete libft (it is not allowed)
