+++ Makefile
@@ -6,7 +6,7 @@
 #    By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/11/21 10:59:06 by mbartos           #+#    #+#              #
-#    Updated: 2024/01/25 15:56:31 by mbartos          ###   ########.fr        #
+#    Updated: 2024/01/26 13:22:46 by mbartos          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -42,7 +42,7 @@ OBJ_PATH =	obj/
 OBJ =		$(SRC:.c=.o)
 OBJS =		$(addprefix $(OBJ_PATH), $(OBJ))
 
-all: libft $(NAME)
+all: $(NAME)
 
 $(OBJ_PATH)%.o :$(SRC_PATH)%.c
 	@$(CC) $(CFLAGS) $(THREADFLAGS) -c $< -o $@ $(INC_LIBFT)
