--- lib/libft/ft_bzero.c
@@ -1,27 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_bzero.c                                         :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/23 13:17:40 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/31 08:37:06 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	ft_bzero(void *s, size_t n)
-{
-	size_t			i;
-	unsigned char	*new_s;
-
-	new_s = (unsigned char *) s;
-	i = 0;
-	while (i < n)
-	{
-		*new_s++ = 0;
-		i++;
-	}
-}
