--- lib/libft/ft_calloc.c
@@ -1,28 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_calloc.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/24 09:21:32 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/27 11:57:15 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	*ft_calloc(size_t count, size_t size)
-{
-	unsigned char	*memory;
-	unsigned int	max_int;
-
-	max_int = 2147483647;
-	if (count > max_int && size > max_int)
-		return (NULL);
-	memory = (unsigned char *) malloc(count * size);
-	if (memory == NULL)
-		return (NULL);
-	ft_bzero(memory, count * size);
-	return ((void *) memory);
-}
