--- lib/libft/ft_int_atoi.c
@@ -1,41 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_int_atoi.c                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/01/25 14:58:19 by mbartos           #+#    #+#             */
-/*   Updated: 2024/01/25 16:00:02 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-int	ft_int_atoi(const char *str)
-{
-	int		i;
-	long	sign;
-	long	number;
-
-	i = 0;
-	sign = 1;
-	number = 0;
-	while (str[i] == ' ' || (str[i] >= '\t' && str[i] <= '\r'))
-		i++;
-	if (str[i] == '-')
-	{
-		sign = -1;
-		i++;
-	}
-	else if (str[i] == '+')
-		i++;
-	else if (!(str[i] >= '0' && str[i] <= '9'))
-		return (-1);
-	while (str[i] >= '0' && str[i] <= '9')
-		number = number * 10 + str[i++] - '0';
-	number = number * sign;
-	if (number <= INT_MAX && number >= 0)
-		return((int) number);
-	return (-1);
-}
