--- lib/libft/ft_isdigit.c
@@ -1,22 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_isdigit.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/10 14:09:17 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/24 10:46:45 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-int	ft_isdigit(int c)
-{
-	if (c > 47 && c < 58)
-	{
-		return (1);
-	}
-	return (0);
-}
