--- lib/libft/ft_isprint.c
@@ -1,22 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_isprint.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/10 14:09:17 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/24 10:46:50 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-int	ft_isprint(int c)
-{
-	if (c >= 32 && c <= 126)
-	{
-		return (1);
-	}
-	return (0);
-}
