--- lib/libft/ft_lstdelone.c
@@ -1,22 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_lstdelone.c                                     :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/26 13:03:49 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/30 12:09:10 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	ft_lstdelone(t_list *lst, void (*del)(void*))
-{
-	if (lst != NULL && del != NULL)
-	{
-		(*del)(lst->content);
-		free(lst);
-	}
-}
