--- lib/libft/ft_lstmap.c
@@ -1,33 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_lstmap.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/26 14:30:44 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/30 11:15:53 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *))
-{
-	t_list	*new;
-	t_list	*first;
-
-	first = NULL;
-	while (lst != NULL)
-	{
-		new = ft_lstnew((*f)(lst->content));
-		if (new == NULL)
-		{
-			ft_lstclear(&first, del);
-			return (NULL);
-		}
-		ft_lstadd_back(&first, new);
-		lst = lst->next;
-	}
-	return (first);
-}
