--- lib/libft/ft_lstnew.c
@@ -1,28 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_lstnew.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/26 09:47:28 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/26 14:20:48 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-t_list	*ft_lstnew(void *content)
-{
-	t_list	*new_node;
-
-	new_node = malloc(sizeof(t_list));
-	if (new_node == NULL)
-		return (NULL);
-	if (content == NULL)
-		new_node->content = NULL;
-	else
-		new_node->content = content;
-	new_node->next = NULL;
-	return (new_node);
-}
