--- lib/libft/ft_lstsize.c
@@ -1,28 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_lstsize.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/26 10:33:17 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/31 08:38:42 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-int	ft_lstsize(t_list *lst)
-{
-	size_t	length;
-
-	if (lst == NULL)
-		return (0);
-	length = 0;
-	while (lst != NULL)
-	{
-		length++;
-		lst = lst->next;
-	}
-	return (length);
-}
