--- lib/libft/ft_memchr.c
@@ -1,29 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_memchr.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/23 13:18:01 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/31 08:36:43 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	*ft_memchr(const void *s, int c, size_t n)
-{
-	size_t			i;
-	unsigned char	*str;
-
-	str = (unsigned char *) s;
-	i = 0;
-	while (i < n)
-	{
-		if (*(str + i) == (unsigned char) c)
-			return ((void *)(str + i));
-		i++;
-	}
-	return (NULL);
-}
