--- lib/libft/ft_memset.c
@@ -1,28 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_memset.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/23 13:19:30 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/31 08:37:58 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	*ft_memset(void *s, int c, size_t n)
-{
-	size_t			i;
-	unsigned char	*new_s;
-
-	new_s = (unsigned char *) s;
-	i = 0;
-	while (i < n)
-	{
-		*new_s++ = (unsigned char) c;
-		i++;
-	}
-	return (s);
-}
