--- lib/libft/ft_putchar_c.c
@@ -1,19 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_putchar_c.c                                     :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/25 11:06:33 by mbartos           #+#    #+#             */
-/*   Updated: 2023/11/01 09:31:57 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	ft_putchar_c(char c, int *ptr_sum)
-{
-	write(1, &c, 1);
-	*ptr_sum += 1;
-}
