--- lib/libft/ft_putendl_fd.c
@@ -1,22 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_putendl_fd.c                                    :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/25 14:02:00 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/25 14:08:05 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	ft_putendl_fd(char *s, int fd)
-{
-	while (*s)
-	{
-		write(fd, s++, 1);
-	}
-	write(fd, "\n", 1);
-}
