--- lib/libft/ft_putnbr_fd.c
@@ -1,34 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_putnbr_fd.c                                     :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/25 14:05:19 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/25 14:08:41 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-void	ft_putnbr_fd(int n, int fd)
-{
-	long int	n_long;
-
-	n_long = n;
-	if (n_long < 0)
-	{
-		n_long = -n_long;
-		ft_putchar_fd('-', fd);
-	}
-	if (n_long > 9)
-	{
-		ft_putnbr_fd(n_long / 10, fd);
-		ft_putnbr_fd(n_long % 10, fd);
-	}
-	else
-	{
-		ft_putchar_fd(n_long + '0', fd);
-	}
-}
