--- lib/libft/ft_strchr.c
@@ -1,31 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_strchr.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/23 13:19:36 by mbartos           #+#    #+#             */
-/*   Updated: 2023/11/28 17:53:19 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-char	*ft_strchr(const char *s, int c)
-{
-	int		i;
-
-	if (!s)
-		return (NULL);
-	i = 0;
-	while (s[i])
-	{
-		if (s[i] == (unsigned char) c)
-			return ((char *) &s[i]);
-		i++;
-	}
-	if ((unsigned char) c == '\0')
-		return ((char *) &s[i]);
-	return (NULL);
-}
