--- lib/libft/ft_strjoin.c
@@ -1,32 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_strjoin.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/24 12:11:04 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/24 13:25:51 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-char	*ft_strjoin(char const *s1, char const *s2)
-{
-	char	*str;
-	char	*str_first;
-	int		str_len;
-
-	str_len = ft_strlen(s1) + ft_strlen(s2);
-	str = (char *) malloc((str_len + 1) * sizeof(char));
-	if (str == NULL)
-		return (NULL);
-	str_first = str;
-	while (*s1)
-		*str++ = *s1++;
-	while (*s2)
-		*str++ = *s2++;
-	*str = '\0';
-	return (str_first);
-}
