--- lib/libft/ft_strlcat.c
@@ -1,33 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_strlcat.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/23 15:36:43 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/25 16:20:46 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-size_t	ft_strlcat(char *dest, const char *src, size_t size)
-{
-	size_t		i;
-	size_t		length_dest;
-	size_t		length_src;
-
-	length_dest = ft_strlen(dest);
-	length_src = ft_strlen(src);
-	if (size <= length_dest)
-		return (length_src + size);
-	i = 0;
-	while (src[i] && (length_dest + i + 1 < size))
-	{
-		dest[length_dest + i] = src[i];
-		i++;
-	}
-	dest[length_dest + i] = '\0';
-	return (length_src + length_dest);
-}
