--- lib/libft/ft_strncmp.c
@@ -1,35 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_strncmp.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/11 15:11:59 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/24 10:48:26 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-int	ft_strncmp(const char *s1, const char *s2, size_t n)
-{
-	unsigned int	i;
-	unsigned char	*str1;
-	unsigned char	*str2;
-
-	str1 = (unsigned char *) s1;
-	str2 = (unsigned char *) s2;
-	i = 0;
-	while ((*str1 != 0 || *str2 != 0) && i < n)
-	{
-		if (*str1 < *str2)
-			return (-1);
-		else if (*str1 > *str2)
-			return (1);
-		str1++;
-		str2++;
-		i++;
-	}
-	return (0);
-}
