--- lib/libft/ft_tolower.c
@@ -1,20 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_tolower.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/23 13:19:58 by mbartos           #+#    #+#             */
-/*   Updated: 2023/10/31 08:35:52 by mbartos          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-int	ft_tolower(int c)
-{
-	if (c >= 65 && c <= 90)
-		c = c + 32;
-	return (c);
-}
