+++ src/philo.c
@@ -12,13 +12,15 @@
 
 #include "philo.h"
 
+#include <stdio.h> //delete!
+
 void *routine(void *program)
 {
 	t_program *program2;
 
 	program2 = (t_program *) program;
 	pthread_mutex_lock(&program2->mutex);
-	ft_printf("Test from thread!\n");
+	printf("Test from thread!\n");
 	pthread_mutex_unlock(&program2->mutex);
 	return (NULL);
 }
@@ -27,25 +29,25 @@ void	ft_print_program_struct(t_program *program)
 {
 	int i;
 
-	ft_printf("========= PROGRAM STRUCT =========\n");
-	ft_printf("Init time to die: %d\n", program->init_time_to_die);
-	ft_printf("Init time to eat: %d\n", program->init_time_to_eat);
-	ft_printf("Init time to sleep: %d\n", program->init_time_to_sleep);
-	ft_printf("Total num of forks(and philosophers): %d\n", program->nof_forks);
-	ft_printf("Actual time: %d\n", program->time);
-	ft_printf("MAX EAT ROUNDS: %d\n", program->max_eat_rounds);
-	ft_printf("\n");
+	printf("========= PROGRAM STRUCT =========\n");
+	printf("Init time to die: %d\n", program->init_time_to_die);
+	printf("Init time to eat: %d\n", program->init_time_to_eat);
+	printf("Init time to sleep: %d\n", program->init_time_to_sleep);
+	printf("Total num of forks(and philosophers): %d\n", program->nof_forks);
+	printf("Actual time: %d\n", program->time);
+	printf("MAX EAT ROUNDS: %d\n", program->max_eat_rounds);
+	printf("\n");
 	i = 0;
 	while (i < program->nof_forks)
 	{
-		ft_printf("========= Philosopher %d =========\n", i);
-		ft_printf("Fork %d on table: %d\n", i, program->forks[i]);
-		ft_printf("Philo %d - time to die: %d\n", i, program->philos_arr[i].time_to_die);
-		ft_printf("Philo %d - time to eat: %d\n", i, program->philos_arr[i].time_to_eat);
-		ft_printf("Philo %d - time to sleep: %d\n", i, program->philos_arr[i].time_to_sleep);
-		ft_printf("Philo %d - left fork in hand: %d\n", i, program->philos_arr[i].left_fork);
-		ft_printf("Philo %d - right fork in hand: %d\n", i, program->philos_arr[i].right_fork);
-		ft_printf("\n");
+		printf("========= Philosopher %d =========\n", i);
+		printf("Fork %d on table: %d\n", i, program->forks[i]);
+		printf("Philo %d - time to die: %d\n", i, program->philos_arr[i].time_to_die);
+		printf("Philo %d - time to eat: %d\n", i, program->philos_arr[i].time_to_eat);
+		printf("Philo %d - time to sleep: %d\n", i, program->philos_arr[i].time_to_sleep);
+		printf("Philo %d - left fork in hand: %d\n", i, program->philos_arr[i].left_fork);
+		printf("Philo %d - right fork in hand: %d\n", i, program->philos_arr[i].right_fork);
+		printf("\n");
 		i++;
 	}
 }
